requirements_full.txt
absl-py==2.1.0
alabaster==1.0.0
alembic==1.14.0
appdirs==1.4.4
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
babel==2.17.0
certifi==2024.12.14
charset-normalizer==3.4.1
clarabel==0.10.0
cloudpickle==3.1.0
colorama==0.4.6
colorlog==6.9.0
colorlover==0.3.0
comm==0.2.2
ConfigSpace==1.2.1
contourpy==1.3.1
cufflinks==0.17.3
cvxpy==1.6.5
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.14
decorator==5.2.1
dill==0.3.9
docutils==0.21.2
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.69.0
gurobipy==12.0.2
gym==0.26.2
gym-notices==0.0.8
h5py==3.12.1
idna==3.10
imagesize==1.4.1
iniconfig==2.3.0
ipykernel==6.29.5
ipython==8.35.0
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.43.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.5.0
mpmath==1.3.0
multiprocess==0.70.17
namex==0.0.8
narwhals==1.42.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.0
opt_einsum==3.4.0
optree==0.13.1
optuna==4.1.0
osqp==1.0.3
packaging==24.2
pandas==2.2.3
pandasgui==0.2.15
parso==0.8.4
pathos==0.3.3
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.7
plotly==6.1.2
pluggy==1.6.0
pmdarima==2.0.4
pox==0.3.5
ppft==1.7.6.9
prompt_toolkit==3.0.51
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
pyarrow==20.0.0
pyepo==0.4.5
Pygments==2.19.1
pymgrid==1.2.2
pynput==1.8.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-microgrid==1.4.1
pytz==2024.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
qtstylish==0.1.5
requests==2.32.3
rich==13.9.4
scikit-base==0.12.0
scikit-learn==1.5.2
scipy==1.15.0
scs==3.2.7.post2
seaborn==0.13.2
shap==0.46.0
six==1.17.0
skforecast==0.14.0
sktime==0.35.0
slicer==0.0.8
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.37
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
utilsforecast==0.2.12
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.1
xgboost==2.1.3
